move remaining stuff to respective domains
