+++ server/src/domains/channels/controllers/http/channel.ts
@@ -1,8 +1,8 @@
 import type { Response } from "express";
-import type { ValidatedRequest } from "../../../types.js";
-import { UserService } from "../../users/index.js";
-import type { ChannelModel } from "../models/channel.js";
-import { getOtherUserProfileInDirectMessageChannelSchema } from "../validators/channel.js";
+import type { ChannelModel } from "../../index.js";
+import type { UserService } from "../../../users/index.js";
+import type { ValidatedRequest } from "../../../../types.js";
+import type { getOtherUserProfileInDirectMessageChannelSchema } from "../../validators/channel.js";
 
 export class ChannelControllerHTTP {
   constructor(
