+++ server/src/domains/channels/index.ts
@@ -1,4 +1,9 @@
-export { ChannelControllerHTTP } from "./controllers/channel.js";
-export { MessageControllerHTTP } from "./controllers/message.js";
+export { ProfileSubscriptionControllerWS } from "../notifications/index.js";
+export { ChannelControllerHTTP } from "./controllers/http/channel.js";
+export { MessageControllerHTTP } from "./controllers/http/message.js";
+export { ChatControllerWS } from "./controllers/ws/chat.js";
 export { ChannelModel } from "./models/channel.js";
 export { MessageService } from "./services/message.js";
+export * from "./validators/channel.js";
+export * from "./validators/message.js";
+
