+++ server/src/domains/notifications/index.ts
@@ -1,0 +1,2 @@
+export { ProfileSubscriptionControllerWS } from "./controllers/ws/profileSubscription.js"
+export { ProfileSubscriptionStore } from "./stores/profileSubscription.js";
