+++ server/src/domains/notifications/services/profileSubscription.ts
@@ -1,0 +1,3 @@
+class ProfileSubscriptionService {
+  constructor() {}
+}
