+++ server/src/domains/relations/index.ts
@@ -1,5 +1,7 @@
+export { RelationControllerHTTP } from "./controllers/relation.js";
 export * from "./entities/invite.js";
 export * from "./entities/relation.js";
-export { RelationControllerHTTP } from "./controllers/relation.js";
 export { FriendInviteModel } from "./models/invite.js";
 export { RelationModel } from "./models/relation.js";
+export { RelationService } from "./services/relation.js";
+
