+++ server/src/domains/users/index.ts
@@ -1,7 +1,9 @@
 export { StatusControllerHTTP } from "./controllers/http/status.js";
 export { UserControllerHTTP } from "./controllers/http/user.js";
 export * from "./entities/user.js";
-export type { User, UserProfile, UserStatus } from "./interfaces/user.js";
+export type * from "./interfaces/user.js";
 export { UserModel } from "./models/user.js";
 export { StatusService } from "./services/status.js";
 export { UserService } from "./services/user.js";
+export * from "./validators/status.js";
+export * from "./validators/user.js";
