--- server/src/stores/index.ts
@@ -1,3 +1,0 @@
-import ProfileSubscriptionStore from "./profileSubscriptionStore.js";
-
-export { ProfileSubscriptionStore };
