--- server/src/websocket/controllers/index.ts
@@ -1,4 +1,0 @@
-import ChatControllerWS from "./chatController.js";
-import ProfileSubscriptionControllerWS from "./profileSubscriptionController.js";
-
-export { ChatControllerWS, ProfileSubscriptionControllerWS };
